ArchivedReportFooter listening directly the reportsAction and getting the closeAction
